Add truncate time option
It turns out, it is possible to have individual series/instances within the
same study to have time-related attributes that both include and not include the fraction.

E.g., "121500.000000" and "121500", which will lead to an undesired result if that
key is used for sorting.

This commit adds a flag that will check for the presence of 0 fraction, and will truncate
it when found for the values of all keys that end in *Time.